% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_theme.R
\name{apa_theme}
\alias{apa_theme}
\title{Format a table using APA style.}
\usage{
apa_theme(ft)
}
\arguments{
\item{ft}{A \code{flextable} or \code{gtsummary} object. Will convert to \code{flextable} if not a \code{flextable} object.}
}
\description{
This function can be applied to \code{gtsummary} and \code{flextable} objects. It will format the table
to be in APA style. You still need to manually add a title and save it as a .docx, though.
}
