% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotProbe.R
\name{ggplotProbe}
\alias{ggplotProbe}
\title{Convert object from \code{plotProbe} into a \code{ggplot} object.}
\usage{
ggplotProbe(model_res, xlim = xlim, mycolor)
}
\arguments{
\item{model_res}{An object produced by \href{https://www.rdocumentation.org/packages/semTools/versions/0.5-6/topics/plotProbe.html#:~:text=335.\%2010.3390/psych3030024-,See\%20Also,latent\%20interaction\%20when\%20the\%20results\%20are\%20obtained\%20from\%20residual\%2Dcentering\%20approach.,-Examples}{\code{probe2WayMC} or other related functions}.}

\item{xlim}{Your specified \code{xlim} parameter. E.g., \code{c(-1, 1)}.}

\item{mycolor}{Specified colors. Can be one or more colors.}
}
\description{
This is an updated/modified version of \href{https://stackoverflow.com/users/13716967/cookesd}{cookesd}'s function,
which can be found \href{https://stackoverflow.com/questions/64931105/plotting-interaction-graph-solely-based-on-slope}{here on StackOverflow}.
}
