% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_alpha.R
\name{mice_alpha}
\alias{mice_alpha}
\title{Calculates Cronbach's alpha with multiply imputed data.}
\usage{
mice_alpha(imp, varlist, bnum, title)
}
\arguments{
\item{imp}{A \code{mids} object.}

\item{varlist}{Variables that make up a specific subscale, e.g., \code{c("x1", "x2", "x3")}.}

\item{bnum}{An integer specifying the number of bootstrap samples to be taken.}

\item{title}{Character, title of your Cronbach's alpha table Optional.}
}
\description{
This function calculates Cronbach's alpha from variables in a \code{mids} object.
#' Based on Dion Groothof's \href{https://stackoverflow.com/questions/70816175/how-do-i-calculate-cronbachs-alpha-on-multiply-imputed-data}{StackOverflow post}.
}
\details{
The output will be a \code{flextable} formatted in APA style.
}
