% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_df.R
\name{mice_df}
\alias{mice_df}
\title{Create a descriptive statistics table from imputed data.}
\usage{
mice_df(imp, vs, title, nm)
}
\arguments{
\item{imp}{A \code{mids} object.}

\item{vs}{Character vector, variables from \code{imp}. E.g., \code{c("bmi", "chl")}.}

\item{title}{Character vector, title of your correlation matrix. Optional.}

\item{nm}{Character vector, preferred variable names. E.g., \code{c("BMI", "Cholesterol")}. Optional.}
}
\description{
This takes data from a \code{mids} object and turns the variables you specify
into table of descriptive statistics This is useful when handling imputed data.
}
\details{
The output will be a \code{flextable} formatted in APA style.
}
