% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lavaan_mi.R
\name{tidy_lavaan_mi}
\alias{tidy_lavaan_mi}
\title{Tidy a \code{lavaan.mi} object.}
\usage{
tidy_lavaan_mi(x)
}
\arguments{
\item{x}{A \code{lavaan.mi} object, such as those returned from \code{\link[semTools:runMI]{semTools::lavaan.mi()}}.}
}
\description{
Put description here later.
}
